move _set_mpl_backend tests to test_helpers
